pyproject.toml
[project]
name = "pyreports"
version = "0.1.0"
authors = [
    { name = "Matt Reimer", email = "matt@northarrowresearch.com" }
]
description = "Client scripts for the Riverscapes Reporting API"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
  "apsw>=3.49.1.0",
  "geopandas>=1.0.1",
  "graphql-core==3.2.6",
  "lxml==5.3.1",
  "matplotlib>=3.9.4",
  "pyarrow>=21.0.0",
  "jsonschema>=4.17.3",
  "python-dateutil==2.9.0.post0",
  "requests==2.32.3",
  "inquirer==3.4.0",
  "rsxml==2.0.6",
  "semver==3.0.4",
  "setuptools==76.0.0",
  "six==1.17.0",
  "termcolor==2.5.0",
  "urllib3>=2.3",
  "riverscapes-metadata @ git+https://github.com/Riverscapes/RiverscapesXML.git@master#subdirectory=riverscapes_metadata",
  "pandas>=2.3.3",
  "tqdm>=4.67.1",
  "questionary>=2.1.1",
]

[project.optional-dependencies]
geo = [
  "gdal==3.11.4",
  "shapely>=2.1.1",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
pyreports = ["**/*.graphql", "**/*.json"]

[tool.ruff]
line-length = 240
target-version = "py312"

[tool.ruff.format]
quote-style = "preserve"

[tool.ruff.lint]
# I = isort (import sorting fix)
# E, F = Flake8 (standard errors)
# PL = Pylint (replaces Pylint's rules)
# N = pep8 naming checks
select = ["I", "E", "F", "PL", "N"]

[tool.ruff.lint.isort]
# Tells Ruff exactly which imports are "local" to your project
known-first-party = ["pyreports"]
combine-as-imports = true

[dependency-groups]
dev = [
    "pylint>=4.0.5",
]
